Shared logic for tasks and custom facts
* Shared library at `lib/os_patching/os_patching.rb`
* Extract all sharable code from tasks (`tasks/*.rb`) and move to shared library
* Move remaining task code to `tasks.real/*.rb` since tasks cannot reference
  shared libraries yet.
* Add an additional `Makefile` at `tasks/Makefile` to concatenate the shared
  library and task files together to provide files that puppet can use. You
  must run `cd tasks && make` each time to update the files before building
  a module release until Puppet Enterprise addresses this in the tasks product
* Remove all calls to `facter`/`puppet facts` and use the Ruby API instead, via
  shared library (fixes #68)
* Rewrite `output()` function to iterate the passed in hash in order to allow
  single function to work for all tasks
* Ensure all calls to external programs have a timeout (10 seconds), slow
  running tasks (updates, cache cleans, fact refreshes) will use any supplied
  timeout or default to 3600 seconds (fixes #94)
* Rewrite the custom fact code to obtain all data by calling a single function
  in the shared library
* Add a `timeout` parameter to the `clean_cache` and `refresh_fact` tasks
* Add a `debug` parameter to all tasks, Support writing to STDOUT when debug
  mode is enabled
* All big-ticket jobs get their own reusable function in the shared library (eg
  clean cache, update, refresh facts, etc)
* Extract in-lined code to shared functions to support:
  * `windows?` - Detecting Windows
  * `os_patching_dir` - Detecting state directory
  * `chunk_*` - Progressive fact generation
  * `history` - Writing the history file
  * `output` - Outputting error result
  * `err` - Outputting task error
  * `run_with_timeout` - Making system calls
  * `fact` - Generate whole fact
  * `get_logger` - Get logging instance
  * `get_params` - Parse params
  * `clean_cache` - Clean the cache
  * `refresh_facts` - Refresh the facts
  * `supported_platform` - Check we are running on a supported platform
  * `get_timeout` - Get any timeout requested by user or switch to default
* Document all functions in shared library (YARD/puppet strings)
