+++ tasks/clean_cache.json
@@ -4,5 +4,13 @@
   "supports_noop": false,
   "description": "Clean patch caches (yum/dpkg) via a task",
   "parameters": {
+    "debug": {
+      "description": "Print debugging information",
+      "type": "Optional[Boolean]"
+    },
+    "timeout": {
+      "description": "How many seconds should we wait until timing out the patch run? (Defaults to 3600 seconds)",
+      "type": "Optional[Integer]"
+    }
   }
 }
