+++ tasks/refresh_fact.json
@@ -4,5 +4,13 @@
   "supports_noop": false,
   "description": "Force a refresh of the os_patching fact cache via a task",
   "parameters": {
+    "debug": {
+      "description": "Print debugging information",
+      "type": "Optional[Boolean]"
+    },
+    "timeout": {
+      "description": "How many seconds should we wait until timing out the task run? (Defaults to 3600 seconds)",
+      "type": "Optional[Integer]"
+    }
   }
 }
